Sticky tape fell off, that's what happens when you don't buy the real thing...
git-svn-id: file:///svn/phpbb/trunk@1324 89ea8834-ac86-4346-8a33-228a782c2dd0

+++ phpBB/posting.php
@@ -167,22 +167,24 @@ function remove_common($percent, $word_id_list = array())
 				}
 			}
 
-			$sql = "DELETE FROM " . SEARCH_WORD_TABLE . "  
-				WHERE word_id IN ($word_id_sql)";
-			$result = $db->sql_query($sql); 
-			if( !$result )
+			if( $word_id_sql != "" )
 			{
-				message_die(GENERAL_ERROR, "Couldn't delete word list entry", "", __LINE__, __FILE__, $sql);
-			}
+				$sql = "DELETE FROM " . SEARCH_WORD_TABLE . "  
+					WHERE word_id IN ($word_id_sql)";
+				$result = $db->sql_query($sql); 
+				if( !$result )
+				{
+					message_die(GENERAL_ERROR, "Couldn't delete word list entry", "", __LINE__, __FILE__, $sql);
+				}
 
-			$sql = "DELETE FROM " . SEARCH_MATCH_TABLE . " 
-				WHERE word_id IN ($word_id_sql)";
-			$result = $db->sql_query($sql); 
-			if( !$result )
-			{
-				message_die(GENERAL_ERROR, "Couldn't delete word match entry", "", __LINE__, __FILE__, $sql);
+				$sql = "DELETE FROM " . SEARCH_MATCH_TABLE . " 
+					WHERE word_id IN ($word_id_sql)";
+				$result = $db->sql_query($sql); 
+				if( !$result )
+				{
+					message_die(GENERAL_ERROR, "Couldn't delete word match entry", "", __LINE__, __FILE__, $sql);
+				}
 			}
-		
 		}
 	}
